chore: remove circular dependency
