+++ projects/cdk/utils/di/inject-form-field.ts
@@ -1,6 +1,6 @@
 import {computed, inject, type InjectOptions, type Signal} from '@angular/core';
 import {NgControl} from '@angular/forms';
-import {type TuiCompatValidationError} from '@taiga-ui/cdk/classes';
+import {type TuiCompatValidationError} from '@taiga-ui/cdk/types';
 
 /**
  * @deprecated use inject(FORM_FIELD) from `@angular/forms/signals` if Angular >= 21
